+++ src/ResourceManager/AzureBatch/Commands.Batch.Test/Pools/StopBatchPoolResizeCommandTests.cs
@@ -34,7 +34,7 @@ namespace Microsoft.Azure.Commands.Batch.Test.Pools
 
         public StopBatchPoolResizeCommandTests(Xunit.Abstractions.ITestOutputHelper output)
         {
-            ServiceManagemenet.Common.Models.XunitTracingInterceptor.AddToContext(new ServiceManagemenet.Common.Models.XunitTracingInterceptor(output));
+            ServiceManagement.Common.Models.XunitTracingInterceptor.AddToContext(new ServiceManagement.Common.Models.XunitTracingInterceptor(output));
             batchClientMock = new Mock<BatchClient>();
             commandRuntimeMock = new Mock<ICommandRuntime>();
             cmdlet = new StopBatchPoolResizeCommand()