+++ src/ResourceManager/HDInsight/Commands.HDInsight.Test/UnitTests/HttpTests.cs
@@ -31,7 +31,7 @@ namespace Microsoft.Azure.Commands.HDInsight.Test
 
         public HttpTests(Xunit.Abstractions.ITestOutputHelper output)
         {
-            ServiceManagemenet.Common.Models.XunitTracingInterceptor.AddToContext(new ServiceManagemenet.Common.Models.XunitTracingInterceptor(output));
+            ServiceManagement.Common.Models.XunitTracingInterceptor.AddToContext(new ServiceManagement.Common.Models.XunitTracingInterceptor(output));
             base.SetupTestsForManagement();
             _httpCred = new PSCredential("hadoopuser", string.Format("Password1!").ConvertToSecureString());