+++ src/ResourceManager/Sql/Commands.Sql.Test/ScenarioTests/ManagedDatabaseCrudScenarioTests.cs
@@ -14,7 +14,7 @@
 
 using Microsoft.Azure.Commands.ScenarioTest.SqlTests;
 using Microsoft.Azure.Management.Sql;
-using Microsoft.Azure.ServiceManagemenet.Common.Models;
+using Microsoft.Azure.ServiceManagement.Common.Models;
 using Microsoft.WindowsAzure.Commands.ScenarioTest;
 using Xunit;
 using Xunit.Abstractions;