+++ src/ResourceManager/Sql/Commands.Sql/Properties/Resources.Designer.cs
@@ -88,15 +88,6 @@ namespace Microsoft.Azure.Commands.Sql.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The action group &apos;AUDIT_CHANGE_GROUP&apos; is not supported. It will be removed in a future release..
-        /// </summary>
-        internal static string auditChangeGroupDeprecationMessage {
-            get {
-                return ResourceManager.GetString("auditChangeGroupDeprecationMessage", resourceCulture);
-            }
-        }
-        
-        /// <summary>
         ///   Looks up a localized string similar to In order to enable Threat Detection, please enable database auditing..
         /// </summary>
         internal static string AuditingIsTurnedOff {